Adding projectile sub-components to a seprate file, next will be to replace the `_area()` functions in ProjectileModal.jsx with only the elememt tags. Weeeooooo
